+++ indra/newview/llviewermessage.cpp
@@ -1759,7 +1759,7 @@ void inventory_offer_handler(LLOfferInfo* info)
 
 	// Strip any SLURL from the message display. (DEV-2754)
 	std::string msg = info->mDesc;
-	int indx = msg.find(" ( http://slurl.com/secondlife/");
+	size_t indx = msg.find(" ( http://slurl.com/secondlife/");
 	if (indx == std::string::npos)
 	{
 		// try to find new slurl host
